Move dev bundle configuration to extension
